DRY up plugins
